+++ libafl_frida/src/asan/mod.rs
@@ -1,3 +1,3 @@
-pub mod asan_errors;
-pub mod asan_hook_funcs;
+pub mod errors;
+pub mod hook_funcs;
 pub mod asan_rt;
